Extracted common behavior of PathQuery and PathsQuery into AbstractPathQuery
